+++ apis/core/bootstrap/organizations.ts
@@ -1,5 +1,5 @@
 import { turtle } from '@tpluscode/rdf-string'
-import { hydra, rdf, rdfs, schema } from '@tpluscode/rdf-ns-builders'
+import { dcat, hydra, rdf, rdfs, schema, _void } from '@tpluscode/rdf-ns-builders'
 import { cc } from '@cube-creator/core/namespace'
 
 export const organizations = turtle`
@@ -17,10 +17,13 @@ export const organizations = turtle`
     a ${schema.Organization} ;
     ${cc.publishGraph}  <https://lindas.admin.ch/foen/cube> ;
     ${cc.namespace} <https://environment.ld.admin.ch/foen/> ;
+    ${schema.dataset} <https://environment.ld.admin.ch/.well-known/void> ;
     ${rdfs.label} "Bundesamt für Umwelt BAFU"@de ;
     ${rdfs.label} "Office fédéral de l'environnement OFEV"@fr ;
     ${rdfs.label} "Ufficio federale dell'ambiente UFAM"@it ;
     ${rdfs.label} "Federal Office for the Environment FOEN"@en ;
+    ${dcat.accessURL} <https://environment.ld.admin.ch/query> ;
+    ${_void.sparqlEndpoint} <https://environment.ld.admin.ch/sparql> ;
   .
 }
 `