+++ cli/lib/variables.ts
@@ -2,6 +2,7 @@ import { Literal, NamedNode, Term } from 'rdf-js'
 import { HydraClient } from 'alcaeus/alcaeus'
 import * as Csvw from '@rdfine/csvw'
 import DatasetExt from 'rdf-ext/lib/Dataset'
+import { PublishJob } from '../../packages/model/Job'
 
 declare module 'barnard59-core/lib/Pipeline' {
   interface VariableNames {
@@ -15,6 +16,7 @@ declare module 'barnard59-core/lib/Pipeline' {
     'publish-graph-store-endpoint': string
     'publish-graph-store-user': string
     'publish-graph-store-password': string
+    'publish-job': PublishJob
     'target-graph': string
     revision: number
     namespace: string